\name{cadcomplete}
\alias{cadcomplete}
\alias{cadincomplete}
\docType{data}
\title{Coronary artery disease data}
\description{
Coronary artery disease data
}
\usage{data(cadcomplete)}
\format{
  A data frame with 236 observations on the following 14 variables.
  \describe{
    \item{Sex}{a numeric vector}
    \item{AngPec}{a numeric vector}
    \item{AMI}{a numeric vector}
    \item{QWave}{a numeric vector}
    \item{QWavecode}{a numeric vector}
    \item{STcode}{a numeric vector}
    \item{STchange}{a numeric vector}
    \item{SuffHeartF}{a numeric vector}
    \item{Hypertrophi}{a numeric vector}
    \item{Hyperchol}{a numeric vector}
    \item{Smoker}{a numeric vector}
    \item{Inherit}{a numeric vector}
    \item{Heartfail}{a numeric vector}
    \item{CAD}{a numeric vector}
  }
}
\details{

}
\source{
  ~~ reference to a publication or URL from which the data were obtained ~~
}
\references{
  ~~ possibly secondary sources and usages ~~
}
\examples{
data(cadcomplete)
## maybe str(cadcomplete) ; plot(cadcomplete) ...
}
\keyword{datasets}
