\name{carcass}
\alias{carcass}
\non_function{}
\title{lean meat contents of pig carcasses}
\usage{data(carcass)}
\description{
  Measurement of lean meat percentage and thickness of meat and fat
  layers at different locations of 344 pig carcasses.
}
\format{
  A data frame with 7 variables and 344 observations. LMP is the lean
  meat percentage determined by dissection. F11, F12 and F13 are
  measurements of fat depth at three different locations on the carcass
  and M11, M12 and M13 are measurements of meat depth at the same
  locations.
}
\source{
 Busk, Hans XXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXX
}
\references{
~~ possibly secondary sources and usages ~~
}
\keyword{datasets}
