\name{edges-methods}
\docType{methods}
\alias{edges-methods}
\alias{nodes-methods}
\alias{edges}
\alias{nodes}
\alias{edges,mim,ANY-method}
\alias{nodes,mim-method}
\title{Retrieve edges and nodes of a mim model}
\description{
  Retrieve edges and nodes of a mim model
}

\section{Methods}{
\describe{
  \item{edges}{Return the edges of a mim model object}
  \item{nodes}{Return the nodes (vertices) of a mim model object}
}}


\keyword{methods}
%%\keyword{ ~~ other possible keyword(s)}
