\name{imputeMissing}
\alias{imputeMissing}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Impute missing values}
\description{
  Imputes missing values in a data set in MIM after a model has been
  fitted.
  }
\usage{
imputeMissing()
}
%- maybe also 'usage' for other objects documented here.
\value{
  None
}
\references{ David Edwards, An Introduction to Graphical Modelling,
  Springer Verlag, 2002}
\author{Søren Højsgaard, sorenh@agrsci.dk}
\note{Before using mimR, make sure that the MIM program is runnning.}
\seealso{\code{\link{fit.mim}}}
%\examples{
%}
\keyword{models}% at least one, from doc/KEYWORDS

