\name{internal}

%% nameTable-2006.R
\alias{makeNameTable}
\alias{makeNameTable.data.frame}
\alias{makeNameTable.table}
\alias{makeNameTablePrimitive}

%% internal-2006.R
\alias{src2tgt}
\alias{src2tgt.character}
\alias{src2tgt.list}
\alias{src2tgt.NULL}
\alias{string2listLetters}
\alias{string2listNames}
\alias{letters2names}
\alias{names2letters}
\alias{list2stringLetters}
\alias{list2stringNames}
\alias{mimFormulaLetters}
\alias{mimFormulaNames}

\alias{create.table}

\alias{print.mimproperties}
\alias{print.mimsummary}


\title{Internal functions for the mimR package}
\description{These functions should not be called directly by the user}
\keyword{internal}% at least one, from doc/KEYWORDS
