\name{math}
\alias{math}
\non_function{}
\title{Mathematics marks for students}
\usage{data(math)}
\description{
The \code{math} data frame has 88 rows and 5 columns.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{me}{a numeric vector}
    \item{ve}{a numeric vector}
    \item{al}{a numeric vector}
    \item{an}{a numeric vector}
    \item{st}{a numeric vector}
  }
}
\references{ David Edwards, An Introduction to Graphical Modelling,
  Second Edition, Springer Verlag, 2000}
\author{Søren Højsgaard, sorenh@agrsci.dk}
%\note{Before using mimR, make sure that the MIM program is runnning.}
\examples{
data(math)
## maybe str(math) ; plot(math) ...
}
\keyword{datasets}
