\name{mim-class}
\docType{class}
\alias{mim-class}

\title{Class "mim" -- mim model objects }
\description{A mim model object holds a mim model which includes
  log-linear models for contingency tables, covariance selection models
  for the multivariate normal distribution and mixed interaction models
  (a combination of the two former).}
\section{Objects from the Class}{A virtual Class: No objects may be created from it.}
\section{Extends}{
Class \code{"\linkS4class{oldClass}"}, directly.
}
\section{Methods}{
No methods defined with class "bayesnet" in the signature.
}
\references{ ~put references to the literature/web site here ~ }
\author{ ~~who you are~~ }
\note{ ~~further notes~~ }

% ~Make other sections like Warning with \section{Warning }{....} ~

% \seealso{
% 	~~objects to See Also as \code{\link{~~fun~~}}, ~~~
% 	or \code{\linkS4class{CLASSNAME}} for links to other classes
% }
\examples{
##---- Should be DIRECTLY executable !! ----
}
\keyword{classes}




