\name{edges-methods}
\docType{methods}
\alias{edges-methods}
\alias{nodes-methods}
\alias{edges}
\alias{nodes}
\alias{edges,mim,ANY-method}
\alias{nodes,mim-method}
\title{Retrieve edges and nodes of a mim model}
\description{
  Retrieve edges and nodes of a mim model
}
%\usage{
%\S4method{edges}{mim}(object,which)
%\S4method{nodes}{mim}(object,which)
%}
%- maybe also 'usage' for other objects documented here.
%\arguments{
%  \item{object}{A MIM model object}
%  \item{which}{Not used}
%}

\section{Methods}{
\describe{
  \item{edges}{Return the edges of a mim model object}
  \item{nodes}{Return the nodes (vertices) of a mim model object}
}}


\keyword{methods}
%%\keyword{ ~~ other possible keyword(s)}
