\name{mimR}
\alias{mimR}
\title{
The package `mimR': summary information
}
\description{
This package provides an interface to the MIM program for inference in
mixed graphical models.
}
\section{Details}{
\itemize{
  \item mimR is available on Windows platforms only.
  \item mimR requires that the MIM program is installed on the computer.
  \item MIM can be downloaded from http://www.hypergraph.dk.
  \item mimR requires that the R package RDCOMClient package is installed.
  \item The RDCOMClient package can be downloaded from http://www.omegahat.org/RDCOMClient/
  \item For information about required versions of R, MIM and the
  RDCOMClient, please see the information presented when mimR is loaded.
  }
The package is intended as a contribution to the gR-project 
derscribed by Lauritzen (2002).
}
\section{Authors}{
Søren Højsgaard, Biometry Research Unit, Danish Institute of
Agricultural Sciences, DK-8830 Tjele, Denmark
}
\section{Acknowledgements}{
  Thanks to David Edwards, the creator of the MIM program.
}
\references{
Lauritzen, S. L. (2002). gRaphical Models in R.
\emph{R News}, 3(2)39.
}
\keyword{multivariate}
\keyword{models}
\keyword{graphs}
